requirements.txt.txt
pandas
numpy
matplotlib
seaborn
tqdm
openai
anthropic
google-generativeai

